tools: stricter eslint rule for globals
